[Fix] cleaning code
